refactor: index파일을 생성하여 모듈 통합
